clusterWidget: create a separated xmlclusterparser
